refactor(harvest): Extract cron logic into dedicated helper ✨
