+++ packages/cozy-harvest-lib/test/helpers/triggers.spec.js
@@ -1,10 +1,6 @@
 /* eslint-env jest */
 
-import {
-  buildCronFromFrequency,
-  buildCron,
-  buildAttributes
-} from 'helpers/triggers'
+import { buildAttributes } from 'helpers/triggers'
 
 describe('Triggers Helper', () => {
   describe('buildAttributes', () => {
@@ -48,80 +44,6 @@ describe('Triggers Helper', () => {
           konnector: 'konnectest'
         }
       })
-    })
-  })
-
-  describe('buildCronFromFrequency', () => {
-    const options = {
-      dayOfMonth: 25,
-      dayOfWeek: 4,
-      hours: 14,
-      minutes: 15
-    }
-
-    it('creates default cron (weekly)', () => {
-      expect(buildCronFromFrequency()).toEqual('0 0 0 * * 1')
-    })
-
-    it('creates weekly cron', () => {
-      expect(buildCronFromFrequency('weekly', options)).toEqual('0 15 14 * * 4')
-    })
-
-    it('creates monthly cron', () => {
-      expect(buildCronFromFrequency('monthly', options)).toEqual(
-        '0 15 14 25 * *'
-      )
-    })
-
-    it('creates daily cron', () => {
-      expect(buildCronFromFrequency('daily', options)).toEqual('0 15 14 * * *')
-    })
-
-    it('creates hourly cron', () => {
-      expect(buildCronFromFrequency('hourly', options)).toEqual('0 15 * * * *')
-    })
-  })
-
-  describe('buildCron', () => {
-    const randomDayTimeMock = jest.fn()
-
-    beforeEach(() => {
-      randomDayTimeMock.mockImplementation((min, max) => ({
-        hours: max - 1,
-        minutes: 59
-      }))
-    })
-
-    afterEach(() => {
-      randomDayTimeMock.mockReset()
-    })
-
-    it('returns expected default cron', () => {
-      const konnector = {}
-      const date = new Date('2019-02-07T14:12:00')
-      expect(buildCron(konnector, date, randomDayTimeMock)).toEqual(
-        `0 59 4 * * 4`
-      )
-    })
-
-    it('returns expected monthly cron', () => {
-      const konnector = {
-        frequency: 'monthly'
-      }
-      const date = new Date('2019-02-07T14:12:00')
-      expect(buildCron(konnector, date, randomDayTimeMock)).toEqual(
-        `0 59 4 7 * *`
-      )
-    })
-
-    it('returns expected cron with time interval', () => {
-      const konnector = {
-        time_interval: [0, 12]
-      }
-      const date = new Date('2019-02-07T14:12:00')
-      expect(buildCron(konnector, date, randomDayTimeMock)).toEqual(
-        `0 59 11 * * 4`
-      )
     })
   })
 })
